Moved old ini related code to separate files
